+++ src/sys/ptrace/linux.rs
@@ -26,11 +26,14 @@ cfg_if! {
     }
 }
 
-libc_enum!{
-    #[cfg_attr(not(any(target_env = "musl", target_os = "android")), repr(u32))]
-    #[cfg_attr(any(target_env = "musl", target_os = "android"), repr(i32))]
+#[cfg(not(any(target_env = "musl", target_os = "android")))]
+type type_of_request = u32;
+#[cfg(any(target_env = "musl", target_os = "android"))]
+type type_of_request = i32;
+
+libc_enum! {
     /// Ptrace Request enum defining the action to be taken.
-    pub enum Request {
+    pub enum Request: type_of_request {
         PTRACE_TRACEME,
         PTRACE_PEEKTEXT,
         PTRACE_PEEKDATA,
@@ -109,12 +112,11 @@ libc_enum!{
     }
 }
 
-libc_enum!{
-    #[repr(i32)]
+libc_enum! {
     /// Using the ptrace options the tracer can configure the tracee to stop
     /// at certain events. This enum is used to define those events as defined
     /// in `man ptrace`.
-    pub enum Event {
+    pub enum Event: i32 {
         /// Event that stops before a return from fork or clone.
         PTRACE_EVENT_FORK,
         /// Event that stops before a return from vfork or clone.